Finishes v2
command logic is now handled by the server. There are both regular commands and
the special !spk command. Furthermore, there are now multiple game engine
objects, one per zone. Player login is also handled by the sever, though there
isn't a way to log out. There are still tests to be run and bugs to be fixed
but I've decided to take a detour to enhance the "tool" logic, so expect a v3
next time.