Upgrades to initial v3
In v2, the logic was that each game object would update their own state via LLM
In v3, the logic is that the game engine object of each subzone has the power
to update the game objects within that zone. Only player characters move between
zones, through the "move" handler (whether explictly via a !move command or
implicitly via the !spk command changing the "subzone" field of the character
object).
As it is, every game object only updates themselves. And the tool set is not
well-curated nor well-done. And the game engine cannot propose new tools as in
v2 (the game objects in v2 could propose their tools). So next updates should
see a "target" being added to the functions in the tools. And the capability
for the game engine to propose new tools and mark them for human review.
One very important point about this v3 version is that tools are not executed
in an eval() or in a restricted Lua/python environment (v2 used eval). Instead,
tools are interpreted by LLM calls into updates for the state dictionaries of
each gameobject within a subzone. So the LLM "executes" the tool code. This
opens the possibility for tool which are specified in English, or a mixture of
English and python. It also means there's performance improvements to be made
by paralelizing these calls (should be straightforward with asyncio).
That said, the code is not well-documented and has some unused parts. I am
hoping for a prototype to be done first so we can check the core gameplay and
decide if this approach is fun before overhauling the codebase.